Moved more text functions to text module
