+++ src/molecule/text.py
@@ -2,6 +2,13 @@
 import re
 
 
+def camelize(string):
+    """Format string as camel-case."""
+    # NOTE(retr0h): Taken from jpvanhal/inflection
+    # https://github.com/jpvanhal/inflection
+    return re.sub(r"(?:^|_)(.)", lambda m: m.group(1).upper(), string)
+
+
 def chomp(text: str) -> str:
     """Remove any training spaces from string."""
     return "\n".join([x.rstrip() for x in text.splitlines()])
@@ -28,3 +35,19 @@ def strip_ansi_color(data):
     invisible_codes = re.compile(r"\x1b\[\d*m")
 
     return re.sub(invisible_codes, "", data)
+
+
+def underscore(string):
+    """Format string to underlined notation."""
+    # NOTE(retr0h): Taken from jpvanhal/inflection
+    # https://github.com/jpvanhal/inflection
+    string = re.sub(r"([A-Z]+)([A-Z][a-z])", r"\1_\2", string)
+    string = re.sub(r"([a-z\d])([A-Z])", r"\1_\2", string)
+    string = string.replace("-", "_")
+
+    return string.lower()
+
+
+def title(word: str) -> str:
+    """Format title."""
+    return " ".join(x.capitalize() or "_" for x in word.split("_"))
